rename files and directories
