+++ src/pages/workspaces/workspace/SubmissionHistory.js
@@ -19,7 +19,7 @@ import * as Nav from 'src/libs/nav';
 import { forwardRefWithName, useCancellation, useOnMount } from 'src/libs/react-utils';
 import * as Style from 'src/libs/style';
 import * as Utils from 'src/libs/utils';
-import UpdateUserCommentModal from 'src/pages/workspaces/workspace/jobHistory/UpdateUserCommentModal';
+import UpdateUserCommentModal from 'src/pages/workspaces/workspace/submissionHistory/UpdateUserCommentModal';
 import { rerunFailures } from 'src/pages/workspaces/workspace/workflows/FailureRerunner';
 import { wrapWorkspace } from 'src/workspaces/container/WorkspaceContainer';
