+++ src/pages/workspaces/workspace/submissionHistory/WorkflowDashboard.js
@@ -25,7 +25,7 @@ import Events, { extractWorkspaceDetails } from 'src/libs/events';
 import { forwardRefWithName, useCancellation, useOnMount } from 'src/libs/react-utils';
 import * as Style from 'src/libs/style';
 import * as Utils from 'src/libs/utils';
-import CallTable from 'src/pages/workspaces/workspace/jobHistory/CallTable';
+import CallTable from 'src/pages/workspaces/workspace/submissionHistory/CallTable';
 import { UriViewer } from 'src/workspace-data/data-table/uri-viewer/UriViewer';
 import { wrapWorkspace } from 'src/workspaces/container/WorkspaceContainer';
